+++ e2e/packages/sync-test/data/expectClientData.ts
@@ -1,7 +1,8 @@
 import { Page, expect } from "@playwright/test";
 import { Data } from "./types";
-import { readClientStore } from "./readClientStore";
 import config from "../../contracts/mud.config";
+import { readComponentValue } from "./readComponentValue";
+import { encodeEntity } from "@latticexyz/store-sync/recs";
 
 /**
  * Confirms that the client state equals the given state by reading from the client's data store
@@ -9,7 +10,7 @@ import config from "../../contracts/mud.config";
 export async function expectClientData(page: Page, data: Data) {
   for (const [table, records] of Object.entries(data)) {
     for (const record of records) {
-      const value = await readClientStore(page, [config.namespace, table, record.key]);
+      const value = await readComponentValue(page, table, encodeEntity(config.tables[table].keySchema, record.key));
       expect(value).toEqual(record.value);
     }
   }